Change maxAgeHours from 18 to 4
With maxAgeHours of 18, an early morning flight ended up archiving
many hours of data generated while testing on the previous afternoon.
This meant we wrote the better part of a terabyte of time series data
that had nothing to do with the flight.

Four hours of lookback should be more than enough.